refactor(legacy-tabs): rename tabs to legacytabs
