+++ package.json
@@ -171,6 +171,7 @@
     "react-native": "0.69.10",
     "react-native-appboy-sdk": "1.35.1",
     "react-native-bootsplash": "3.2.0",
+    "react-native-collapsible-tab-view": "6.1.4",
     "react-native-config": "https://github.com/artsy/react-native-config.git#v1.4.12-artsy",
     "react-native-credit-card-input": "0.4.1",
     "react-native-dev-menu-android": "1.0.10",
@@ -187,7 +188,7 @@
     "react-native-linear-gradient": "2.6.2",
     "react-native-localize": "2.0.1",
     "react-native-network-info": "5.2.1",
-    "react-native-pager-view": "6.0.0",
+    "react-native-pager-view": "6.2.0",
     "react-native-push-notification": "8.1.1",
     "react-native-reanimated": "2.13.0",
     "react-native-reanimated-zoom": "0.3.3",
