+++ src/app/Components/LegacyTabs/ContentTabs.tsx
@@ -1,7 +1,7 @@
-import { Tab, TabsProps } from "app/Components/Tabs"
+import { Tab, TabsProps } from "app/Components/LegacyTabs"
 import React, { useState } from "react"
 import { LayoutRectangle } from "react-native"
-import { TabBarContainer } from "./TabBarContainer"
+import { TabBarContainer } from "./LegacyTabBarContainer"
 
 /**
  * Renders a  scrollable list of tabs. Tabs are not evenly spaced across screen
