+++ src/app/Components/StickyTabPage/StickyTabPageTabBar.tsx
@@ -1,7 +1,7 @@
-import { NavigationalTabs } from "app/Components/Tabs"
+import { NavigationalTabs } from "app/Components/LegacyTabs"
+import { useScreenDimensions } from "app/utils/hooks"
 import React, { useEffect, useRef, useState } from "react"
 import { Animated, LayoutRectangle, ScrollView } from "react-native"
-import { useScreenDimensions } from "app/utils/hooks"
 import { useStickyTabPageContext } from "./StickyTabPageContext"
 
 export const StickyTabPageTabBar: React.FC<{
