+++ src/app/Scenes/City/Components/LegacyTabBar.tests.tsx
@@ -1,6 +1,6 @@
-import TabBar from "app/Components/TabBar"
 import { renderWithWrappersLEGACY } from "app/utils/tests/renderWithWrappers"
 import { Animated } from "react-native"
+import TabBar from "./LegacyTabBar"
 
 it("renders without throwing an error", () => {
   renderWithWrappersLEGACY(
