+++ src/app/Scenes/Fair/Fair.tests.tsx
@@ -1,8 +1,8 @@
 import { FairTestsQuery } from "__generated__/FairTestsQuery.graphql"
+import { NavigationalTabs, Tab } from "app/Components/LegacyTabs"
 import { __globalStoreTestUtils__ } from "app/store/GlobalStore"
 import { extractText } from "app/utils/tests/extractText"
 import { renderWithWrappers, renderWithWrappersLEGACY } from "app/utils/tests/renderWithWrappers"
-import { NavigationalTabs, Tab } from "app/Components/Tabs"
 import { graphql, QueryRenderer } from "react-relay"
 import { act } from "react-test-renderer"
 import { useTracking } from "react-tracking"
