+++ src/app/Scenes/Fair/Fair.tsx
@@ -6,8 +6,8 @@ import { ArtworkFilterNavigator, FilterModalMode } from "app/Components/ArtworkF
 import { ArtworkFiltersStoreProvider } from "app/Components/ArtworkFilter/ArtworkFilterStore"
 import { HeaderArtworksFilterWithTotalArtworks as HeaderArtworksFilter } from "app/Components/HeaderArtworksFilter/HeaderArtworksFilterWithTotalArtworks"
 import { HeaderButton } from "app/Components/HeaderButton"
+import { NavigationalTabs, TabsType } from "app/Components/LegacyTabs"
 import { SearchImageHeaderButton } from "app/Components/SearchImageHeaderButton"
-import { NavigationalTabs, TabsType } from "app/Components/Tabs"
 import { goBack } from "app/system/navigation/navigate"
 import { defaultEnvironment } from "app/system/relay/createEnvironment"
 import { useScreenDimensions } from "app/utils/hooks"
